+++ RubyExtension/MSPhysics.rb
@@ -5,23 +5,19 @@
 #   http://sketchucation.com/forums/viewtopic.php?f=323&t=56852
 #
 # Overview
-#   MSPhysics is a physics simulation tool similar to SketchyPhysics. Unlike
-#   SketchyPhysics, MSPhysics has a far more advanced scripting API. When it
-#   came to creating FPS type games in SP, scripters were limited to having
-#   control over user input. Scripters had to add extra work-around code in
-#   order to gain control over mouse input, which was very time consuming and
-#   not as reliable as wanted. MSPhysics, however, gives scripters full control
-#   over all mouse and keyboard events, including the mouse wheel. Instead of
-#   having keyboard keys serve as shortcuts, they can be intercepted and used as
-#   game controls. Instead of having the mouse wheel serve as a shortcut for the
-#   native zoom in/out operation, it can be intercepted and serve as a command
-#   to switch weapons for instance. All such operations might seem a fantasy,
-#   but thanks to Microsoft Windows API, which is heavily implemented in AMS
-#   Library, gaining control over user input is possible. Along the lines,
-#   MSPhysics uses NewtonDynamics 3.14 by Juleo Jerez in order to produce fast
-#   and realistic physics effects. Compared to Newton 1.53, which is used by
-#   SketchyPhysics, Newton 3.14 is faster and more advanced. In many ways the
-#   goal of this project is to bring SketchyPhysics back to life.
+#   MSPhysics is a real-time physics simulation plugin for SketchUp, similar to
+#   SketchyPhysics. Unlike SketchyPhysics, MSPhsyics is a written completely
+#   from scratch, integrating the latest NewtonDynamics physics SDK. The
+#   differences between the two is that MSPhysics is a lot faster, has a far
+#   more advanced scripting API, and comes with a reliable Replay animation
+#   tool. Furthermore MSPhysics uses AMS Library, which on the Windows side,
+#   allows taking control over user input, such as mouse and keyboard, and
+#   switching SketchUp fullscreen. With such advantage, creating FPS type games
+#   in MSPhysics, without interference of various keyboard/mouse shortcuts, is
+#   now a possibility. At the moment, the project is under development,
+#   particularly within the Wiki branch. All in all huge credit goes to Julio
+#   Jerez for writing the NewtonDynamics physics engine. As well as to those who
+#   contributed in making the project a reality.
 #
 # Access
 #   - (Menu) Plugins → MSPhysics → [Option]
@@ -35,13 +31,13 @@
 #   - AMS Library 3.3.0 or later.
 #
 # Version
-#   - MSPhysics 0.9.0
+#   - MSPhysics 0.9.1
 #   - NewtonDynamics 3.14
 #   - SDL 2.0.4
 #   - SDL_mixer 2.0.1
 #
 # Release Date
-#   July 10, 2016
+#   July 26, 2016
 #
 # Licence
 #   MIT © 2015-2016, Anton Synytsia
@@ -49,7 +45,7 @@
 # Credits
 #   - Julio Jerez for the NewtonDynamics physics engine.
 #   - Chris Phillips for ideas from SketchyPhysics.
-#   - István Nagy (PituPhysics) for examples and testing.
+#   - István Nagy (PituPhysics) for testing.
 #
 # ------------------------------------------------------------------------------
 
@@ -77,8 +73,8 @@ if load_me
   module MSPhysics
 
     NAME         = 'MSPhysics'.freeze
-    VERSION      = '0.9.0'.freeze
-    RELEASE_DATE = 'July 10, 2016'.freeze
+    VERSION      = '0.9.1'.freeze
+    RELEASE_DATE = 'July 26, 2016'.freeze
 
     # Create the extension.
     @extension = SketchupExtension.new NAME, 'MSPhysics/main.rb'